chore(components): rename component files to match tag name
