+++ components/src/web-components/visualization/gs-aggregate.tsx
