+++ components/src/web-components/visualization/gs-relative-growth-advantage.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 
-import './relative-growth-advantage-component';
+import './gs-relative-growth-advantage';
 import '../app';
 import { withComponentDocs } from '../../../.storybook/ComponentDocsBlock';
 import { AGGREGATED_ENDPOINT, LAPIS_URL } from '../../constants';
