+++ components/src/web-components/visualization/gs-relative-growth-advantage.tsx
